Added Drop Down Menu
I changed a few things. I made a ShapesMenu.java class which controls the shapes menu, it can open and close with a button. Shapes can be added in functions or as variables, maybe there can be functions in this class that trigger from clicking the screen in Environment.java which clone the given shape and make it draggable. I made some changes to Environment too, some of which are explained in comments. I decided to switch to an AnchorPane, but I'm still not sure this is the best Pane to use or if we should just use Pane() and use absolute coordinates with listeners to change objects positions if the screen is resized. We may want to change is in the future but it works for now.