Update ShapesMenu.java
One solution to this is simply to deselect the element when the menu is hidden. But the client may appreciate knowing the last selected element before hiding menu. (no wrong color if there's no color lol)
As for the color, I don't know where in the JavaFX is the color for the selected cell is. When a cell is deselected, its color should match its selected color (like blue when showing menu again.)
Yuetian doesn't think this is a real issue